refactor: CloseBtn 컴포넌트 분리 및 useModal 파일명 수정
- CloseBtn을 공통으로 쓸 수 있게 컴포넌트 분리
- useModal 파일명 수정
 : Modal.jsx > useModal.jsx
- useModal 디렉토리를 hooks로 이동
