+++ src/components/PostUploader/LocationUploader.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import useModal from './Modal';
+import useModal from '../../hooks/useModal';
 import LocationFinder from './LocationFinder';
 
 const LocationUploader = () => {
